data-ciphers: Integration of cipher negotiation (NCP) (Patch 1)
- The new directive '--data-ciphers algs' has been introduced for RWs with
OpenVPN version 2.5.0 and is needed with OpenVPN version 2.6.x.
This directive negotiates now the best and available cipher algorithm
with the clients. The selection for '--data-ciphers algs' is
between the GCM family and the new CHACHA20-POLY1305. All ciphers
can be combined with another.
The current default is 'AES-256GCM|ChaCha20-Poly1305'

- The new directive '--data-ciphers algs' substitutes '--ncp-disable', therefor
'--ncp-disable' has been removed which fixes the server error with OpenVPN-2.6_dev but
fixes also the deprecation warning for server instances on version 2.5.x .

- New CHACHA-POLY1305 cipher is integrated.

- NCP is now located under the "Advanced Server" section in IPFires ovpnmain.cgi .